+++ src/core/prompts/sections/McpServerUse.ts
@@ -66,7 +66,7 @@ ${
 					return (
 						`## ${server.name} ${config.command ? `(\`${config.command}${config.args && Array.isArray(config.args) ? ` ${config.args.join(" ")}` : ""}\`)` : "- remote server"}` +
 						(tools
-							? `\n\n### ATL - Available Tools List - Always use get_mcp_tool_input_schema before calling a tool that does not have a defined input schema, as it exposes all tool functionality and available parameters such as sorting, ordering, filtering, etc.\n\n${tools}`
+							? `\n\n### ATL - Available Tools List - Always use the build-in tool "get_mcp_tool_input_schema" before calling a tool that does not have a defined input schema after its name and description, as it exposes all tool functionality and available parameters such as sorting, ordering, filtering, etc.\n\n${tools}`
 							: "") +
 						(templates ? `\n\n### RTL - Resource Templates List\n${templates}` : "") +
 						(resources ? `\n\n### DRL - Direct Resources List\n${resources}` : "")